split connect logic into prepare and process
