+++ packages/twenty-server/src/engine/twenty-orm/entity-manager/types/relation-nested-query-fields-by-entity-index.type.ts
@@ -1,15 +1,14 @@
-import { RELATION_NESTED_QUERY_KEYWORDS } from 'twenty-shared/constants';
-
 import {
   ConnectObject,
   DisconnectObject,
 } from 'src/engine/twenty-orm/entity-manager/types/query-deep-partial-entity-with-relation-connect.type';
 
-export type RelationNestedQueryFieldsByEntityIndex = {
+export type RelationConnectQueryFieldsByEntityIndex = {
+  [entityIndex: string]: { [key: string]: ConnectObject };
+};
+
+export type RelationDisconnectQueryFieldsByEntityIndex = {
   [entityIndex: string]: {
-    [RELATION_NESTED_QUERY_KEYWORDS.CONNECT]?: { [key: string]: ConnectObject };
-    [RELATION_NESTED_QUERY_KEYWORDS.DISCONNECT]?: {
-      [key: string]: DisconnectObject;
-    };
+    [key: string]: DisconnectObject;
   };
 };
